+++ HISTORY.md
@@ -57,6 +57,7 @@
 * blocks: made variadic infix inputs collapsible for explicit input lists
 * blocks: added ability to specify empty variadic collapse labels
 * blocks: tweaked variadic slot arrows layout
+* threads: fixed reportAssociativeBool() for empty variadic inputs
 
 ### 2023-05-30
 * blocks: enable explicit input lists in variadic slots, experimental, under construction